tests/requirements-dev.txt
pytest==8.1.1
pytest-mock==3.14.0
requests-mock==1.12.1
